move Linux executeCmd() to libimhex
