+++ lib/libimhex/include/hex/helpers/utils_linux.hpp
@@ -1,0 +1,9 @@
+#pragma once
+
+#if defined(OS_LINUX)
+
+namespace hex {
+    void executeCmd(const std::vector<std::string> &argsVector);
+}
+
+#endif
