+++ main/source/window/linux_window.cpp
@@ -7,6 +7,7 @@
     #include <hex/api/event.hpp>
 
     #include <hex/helpers/utils.hpp>
+    #include <hex/helpers/utils_linux.hpp>
     #include <hex/helpers/logger.hpp>
 
     #include <wolv/utils/core.hpp>
@@ -35,20 +36,6 @@ namespace hex {
             }
         }
         return false;
-    }
-
-    void executeCmd(const std::vector<std::string> &argsVector) {
-        std::vector<char*> cArgsVector;
-        for (const auto &str : argsVector) {
-            cArgsVector.push_back(const_cast<char*>(str.c_str()));
-        }
-        cArgsVector.push_back(nullptr);
-        
-        if (fork() == 0) {
-            execvp(cArgsVector[0], &cArgsVector[0]);
-            log::error("execvp() failed: {}", strerror(errno));
-            exit(EXIT_FAILURE);
-        }
     }
 
     void nativeErrorMessage(const std::string &message) {
